ocr/requirements.txt
# ─────────────────────────────────────────
# Image Processing (OpenCV + Pillow)
# ─────────────────────────────────────────
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

# ─────────────────────────────────────────
# OCR Models
# ─────────────────────────────────────────
transformers==4.46.3          # HuggingFace (TrOCR, GOT-OCR)
torch==2.5.1                  # PyTorch backend
torchvision==0.20.1
accelerate==1.1.1             # HuggingFace accelerate (multi-GPU)
sentencepiece==0.2.0
timm==1.0.11                  # TrOCR vision backbone
paddlepaddle==2.6.2           # PaddleOCR backend
paddleocr==2.9.1

# ─────────────────────────────────────────
# Layout Detection
# ─────────────────────────────────────────
surya-ocr==0.9.1              # Surya layout detection

# ─────────────────────────────────────────
# Post-Processing
# ─────────────────────────────────────────
language-tool-python==2.8.1   # Spelling correction

# ─────────────────────────────────────────
# Fine-Tuning (LoRA — Sprint 6)
# ─────────────────────────────────────────
# peft==0.13.0                # LoRA fine-tuning (uncomment for Sprint 6)
# unsloth==2024.11.0          # Fast LoRA trainer

# ─────────────────────────────────────────
# Storage + Utils
# ─────────────────────────────────────────
minio==7.2.9
redis==5.1.1
structlog==24.4.0
httpx==0.27.2
